Adding project summary and machine learning roles
I have added the project summary and description of the role for machine learning. To my team member, feel free to add or delete something from the summary or the description of the role. If you are confused about what to write, you can follow the structure of ML descriptions.